Backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o .\requirements.txt
blinker==1.8.2
    # via flask
certifi==2024.6.2
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cmdstanpy==1.2.4
    # via prophet
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flask==3.0.3
fonttools==4.53.0
    # via matplotlib
holidays==0.51
    # via prophet
idna==3.7
    # via requests
importlib-resources==6.4.0
    # via prophet
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.0
    # via prophet
numpy==1.26.4
    # via
    #   cmdstanpy
    #   contourpy
    #   matplotlib
    #   pandas
    #   prophet
    #   stanio
packaging==24.1
    # via
    #   matplotlib
    #   plotly
pandas==2.2.2
    # via
    #   cmdstanpy
    #   prophet
pillow==10.3.0
    # via matplotlib
plotly==5.22.0
prophet==1.1.5
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   holidays
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
pytz==2024.1
    # via pandas
requests==2.32.3
sentry-sdk==2.6.0
six==1.16.0
    # via python-dateutil
stanio==0.5.0
    # via cmdstanpy
tenacity==8.4.1
    # via plotly
tqdm==4.66.4
    # via
    #   cmdstanpy
    #   prophet
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   requests
    #   sentry-sdk
werkzeug==3.0.3
    # via flask
